pyproject.toml
[project]
name = "lift-cam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baidu-aip>=4.16.13",
    "chardet>=6.0.0.post1",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.1",
    "redis>=7.2.1",
    "stomp-py>=8.2.0",
]
